Fix tests, number conversions cleanup and remove Util/toString
