+++ src/js/fable-core/DateOffset.ts
@@ -13,9 +13,9 @@
  * Basically; invariant: date.getTime() always return UTC time.
  */
 
-import { create as createDate, daysInMonth, offsetRegex, parseRaw } from "./Date";
+import { create as createDate, dateOffsetToString, daysInMonth, offsetRegex, parseRaw } from "./Date";
 import { fromValue, ticksToUnixEpochMilliseconds, unixEpochMillisecondsToTicks } from "./Long";
-import { compareDates, DateKind, dateOffsetToString, IDateTime, IDateTimeOffset, padWithZeros } from "./Util";
+import { compareDates, DateKind, IDateTime, IDateTimeOffset, padWithZeros } from "./Util";
 
 export default function DateTimeOffset(value: number, offset?: number) {
   const d = new Date(value) as IDateTimeOffset;
